Support for sorting jobs list in yaml
We have encountered an issue that if jobs in view xml output are
not aplhabetically sorted, it's not properly displayed on generated views in
Jenkins.

This might be an issue of old CI version, anyway the fix is general and
can be used widely.

Tests were also updated to check that sorting is working properly on
required places.

Implemented for list and sectioned views where it's possible to define
the `<jobNames>`.